Correction: changes an ID to a class
In the solution provided the third and fourth texts share a class that
sets the font size to 24, however, in my solution, I used ID's
to achieve that same end. Even though the difference (at least visually)
is not perceivable, the last item of the self check list is not
completed if I use my code, as the third element wouldn't have multiple
classes, but rather a class and an ID. Therefore, (and also to avoid
redundancy of the line setting the font to 24) it is important to
perform this commit.